requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
et_xmlfile==2.0.0
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
six==1.17.0
sqlparse==0.5.5
tzdata==2025.3
tzdata2025.3    